Homework 5: Split code into separate files
