+++ course7/homeworks/patrickkostjens/complex/CMakeLists.txt
@@ -2,4 +2,4 @@ cmake_minimum_required(VERSION 3.0)
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -std=c++14")
 
-add_executable(complex complex.cc)
+add_executable(complex main.cc calculator.hh calculator.cc complex.hh complex.cc)
